modules/convert-depth-format/pyproject.toml
[project]
name = "convert-depth-format"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Simon Okutan", email = "simon.okutan@tum.de"},
]

dependencies = [
    "numpy>=1.23.1",
    "pillow==10.3.0",
    "sd-pipeline-typing @ file:///${PROJECT_ROOT}/../sd-pipeline-typing#egg=sd-pipeline-typing",
]

requires-python = "==3.8.*"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = []